% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_addins.rs-package.R
\docType{package}
\name{addins.rs}
\alias{addins.rs}
\title{A Set of RStudio Add-ins}
\description{
Package \strong{\code{addins.rs}} provides a set of RStudio add-ins that
are designed to be used in combination with user-defined RStudio keyboard
shortcuts. These add-ins either insert text at the cursor position (e.g. insert
operators \verb{\%>\%}, \verb{<<-}, \verb{\%$\%}, etc.) or replace symbols
in selected pieces of text, e.g., convert back-slashes to forward-slashes.\cr

\verb{License:} MIT   \cr
\verb{URL:} \url{https://github.com/GegznaV/addins.rs} \cr
\verb{Bug reports and suggestions:} \url{https://github.com/GegznaV/addins.rs/issues} \cr
\verb{Author:} Vilmantas Gegzna
}
\seealso{
Use the links below to learn more about RStudio add-ins
and especially how to use them in combination with user-defined
keyboard shortcuts:\enumerate{

\item{\href{https://rstudio.github.io/rstudioaddins/}{RStudio add-ins};}
\item{\href{https://rstudio.github.io/rstudioaddins/#keyboard-shorcuts}{Keyboard shortcuts};}
\item{\href{https://support.rstudio.com/hc/en-us/articles/206382178-Customizing-Keyboard-Shortcuts}{Customizing keyboard shortcuts.}}
}
}
