% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app--allign-at.R, R/app--set-options.R,
%   R/rs_align_code.R
\name{rs_align_code_at_pattern}
\alias{rs_align_code_at_pattern}
\alias{rs_addins_options}
\alias{align_code}
\alias{rs_align_code}
\alias{rs_align_code_at_equal}
\alias{rs_align_code_at_arrow}
\alias{rs_align_code_at_pipe_operator}
\alias{rs_align_code_at_native_pipe}
\alias{rs_align_code_at_hash}
\alias{rs_align_code_at_vertical_bar}
\title{Align code at certain symbols}
\usage{
rs_align_code_at_pattern(context = rs_get_context())

rs_addins_options(context = rs_get_context())

rs_align_code(
  at_symbol,
  context = rs_get_context(),
  algorithm = c("selection only", "rows in selection")
)

rs_align_code_at_equal(context = rs_get_context())

rs_align_code_at_arrow(context = rs_get_context())

rs_align_code_at_pipe_operator(context = rs_get_context())

rs_align_code_at_native_pipe(context = rs_get_context())

rs_align_code_at_hash(context = rs_get_context())

rs_align_code_at_vertical_bar(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}

\item{at_symbol}{(character) \cr
A regular expression that represents symbol treated as reference for
code alignment.}

\item{algorithm}{(\code{character}) \cr
One of \code{"selection only"} and \code{"rows in selection"}.}
}
\description{
In selected lines, align code at first occurrence of certain symbol in each
row.
}
\note{
Inspired by \url{https://github.com/seasmith/AlignAssign}.
}
