% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{comment}
\alias{comment}
\alias{rs_comment_uncomment}
\title{Comment/Uncomment selected lines}
\usage{
rs_comment_uncomment()
}
\description{
Comment or uncomment selected lines.
}
\concept{'Comment' addins}
