% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_highlight_row_duplicates.R
\name{highlight_row_duplicates}
\alias{highlight_row_duplicates}
\alias{rs_highlight_first_row_duplicates}
\alias{rs_highlight_row_duplicates_all}
\alias{rs_highlight_row_duplicates}
\title{Highlight duplicates of selected rows.}
\usage{
rs_highlight_first_row_duplicates(context = rs_get_context())

rs_highlight_row_duplicates_all(context = rs_get_context())

rs_highlight_row_duplicates(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Functions that highlight duplicates of selected lines:
\itemize{
\item{rs_highlight_first_row_duplicates() duplicates of the first selected
     row (the first row gets deselected);}
\item{rs_highlight_row_duplicates() duplicates of any selected row
     (except the first occurrence);}
\item{rs_highlight_row_duplicates_all() duplicates of any selected row
     (including the first occurrence).}
}
}
