% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_insert_datetime.R
\name{rs_insert_datetime}
\alias{rs_insert_datetime}
\alias{rs_insert_date}
\alias{rs_insert_time}
\title{Insert date and time.}
\usage{
rs_insert_datetime(context = rs_get_context())

rs_insert_date(context = rs_get_context())

rs_insert_time(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
RStudio add-in to insert date and time at the cursor position.
}
