% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_lines.R
\name{insert_line}
\alias{insert_line}
\alias{rs_insert_line}
\alias{rs_insert_line_ss}
\alias{rs_insert_line_ds}
\alias{rs_insert_line_sw}
\title{Insert various lines}
\usage{
rs_insert_line(
  symbol,
  start = rs_get_index_first_selected_col(),
  end = getOption("addins.rs.line_end", 79)
)

rs_insert_line_ss()

rs_insert_line_ds()

rs_insert_line_sw()
}
\arguments{
\item{symbol}{(character) \cr
The symbol that will be repeated to create a line.}

\item{start}{(numeric) \cr
The number of column the line begins at.}

\item{end}{(numeric) \cr
The number of column the line ends at.}
}
\description{
These functions insert sequence of symbols (the line) which by default starts
at the cursor position and ends at symbol (column) 79. (It is recommended
that the line of code would not be longer than 80 symbols.)

Use
\code{rs_insert_line_ss()} to insert \verb{------};\cr
\code{rs_insert_line_ds()} to insert \verb{======};\cr
\code{rs_insert_line_sw()} to insert \verb{~~~~~~} at the cursor position.
}
\seealso{
Other 'Insert at cursor position' addins:
\code{\link{insert_arrows}}
}
\concept{'Insert at cursor position' addins}
