% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_magrittr.R
\name{insert_magrittr_operators}
\alias{insert_magrittr_operators}
\alias{rs_insert_pipe}
\alias{rs_insert_update_pipe}
\alias{rs_insert_exposition_pipe}
\alias{rs_insert_tee_pipe}
\title{Insert magrittr operators}
\usage{
rs_insert_pipe(context = rs_get_context())

rs_insert_update_pipe(context = rs_get_context())

rs_insert_exposition_pipe(context = rs_get_context())

rs_insert_tee_pipe(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
These add-ins insert operators \verb{\%>\%}, \verb{\%<>\%},
\verb{\%$\%},
and \verb{\%T>\%} at the cursor position.
The operators are defined in package \pkg{magrittr}.
}
\note{
Operator does not work unless package \pkg{magrittr} or
equivalent is loaded.
}
\seealso{
\itemize{
\item Forward pipeline operator \link[magrittr:\%>\%]{\%>\%}
\item Compound assignment pipe operator \link[magrittr:\%<>\%]{\%<>\%}
\item Exposition pipe operator \link[magrittr:\%$\%]{\%$\%} and operator \link[base:$]{$}
\item Tee operator \link[magrittr:\%T>\%]{\%T>\%}
}

Other 'Insert at cursor position' add-ins:
\code{\link[=rs_enclose_with_percent]{rs_enclose_with_percent()}},
\code{\link[=rs_insert_infix_in]{rs_insert_infix_in()}},
\code{\link[=rs_insert_integer_division]{rs_insert_integer_division()}},
\code{\link[=rs_insert_matrix_multiplication]{rs_insert_matrix_multiplication()}},
\code{\link[=rs_insert_modulo]{rs_insert_modulo()}}
}
\concept{'Insert at cursor position' add-ins}
