% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_quotes.R
\name{insert_quotes}
\alias{insert_quotes}
\alias{rs_insert_quotes_double_lower}
\alias{rs_insert_quotes_double_upper_c}
\alias{rs_insert_quotes_double_upper_d}
\title{Insert quotes}
\usage{
rs_insert_quotes_double_lower(context = rs_get_context())

rs_insert_quotes_double_upper_c(context = rs_get_context())

rs_insert_quotes_double_upper_d(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
These add-ins insert quotation symbols (e.g., \verb{„}, \verb{“}, and \verb{”})
at the cursor position.
}
\concept{'Insert quotes' add-ins}
