% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_change_path_type.R
\name{path_type}
\alias{path_type}
\alias{rs_to_absolute_path}
\alias{rs_to_relative_path}
\title{Change path type.}
\usage{
rs_to_absolute_path()

rs_to_relative_path()
}
\description{
Change path from absolute to relative and vice versa.
}
