% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_align_code.R
\name{rs_align_code}
\alias{rs_align_code}
\alias{rs_align_code_at_equal}
\alias{rs_align_code_at_arrow}
\alias{rs_align_code_at_pipe}
\title{Align code at certain symbols.}
\usage{
rs_align_code(at_symbol, context = rs_get_context())

rs_align_code_at_equal(context = rs_get_context())

rs_align_code_at_arrow(context = rs_get_context())

rs_align_code_at_pipe(context = rs_get_context())
}
\arguments{
\item{at_symbol}{(character) A regular expression that represents
symbol treated as reference for code alignment.}

\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
Align code of selected lines at first occurance of certain symbol.
}
\note{
Inspired by: \url{https://github.com/seasmith/AlignAssign}
}
