% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enclose.R
\name{rs_enclose_with_percent}
\alias{rs_enclose_with_percent}
\title{Enclose with \%\%}
\usage{
rs_enclose_with_percent(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Call this function as an add-in to enclose selected text with percent operator \verb{\%}.
}
\seealso{
Other 'Insert at cursor position' add-ins:
\code{\link{insert_magrittr_operators}},
\code{\link[=rs_insert_infix_in]{rs_insert_infix_in()}},
\code{\link[=rs_insert_integer_division]{rs_insert_integer_division()}},
\code{\link[=rs_insert_matrix_multiplication]{rs_insert_matrix_multiplication()}},
\code{\link[=rs_insert_modulo]{rs_insert_modulo()}}
}
\concept{'Insert at cursor position' add-ins}
