% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_infix_operator}
\alias{rs_insert_infix_operator}
\title{Insert \%\%.}
\usage{
rs_insert_infix_operator(context = rs_get_context())
}
\arguments{
\item{context}{Object with context of active RStudio document
(class \code{document_context}).}
}
\description{
Call this function as an add-in to insert \code{ \%\% } at the cursor position.
}
\note{
Operator does not work unless library which contains this operator is loaded.
}
\seealso{
Other 'Insert at cursor position' add-ins: \code{\link{insert_magrittr_operators}},
  \code{\link{rs_insert_infix_R}},
  \code{\link{rs_insert_infix_in}},
  \code{\link{rs_insert_infix_not_in}},
  \code{\link{rs_insert_matrix_multiplication}}
}
\concept{'Insert at cursor position' add-ins}
