% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_sort_rows.R
\name{sort_rows}
\alias{sort_rows}
\alias{rs_sort_rows}
\alias{rs_sort_rows_ascending}
\alias{rs_sort_rows_descending}
\title{Sort selected lines.}
\usage{
rs_sort_rows(decreasing = FALSE, numeric.type = "decimal",
  context = rs_get_context())

rs_sort_rows_ascending(context = rs_get_context())

rs_sort_rows_descending(context = rs_get_context())
}
\arguments{
\item{decreasing}{logical.  Should the sort be increasing or
    decreasing? Note that \code{descending=TRUE} reverses the meanings of
    \code{na.last} and \code{blanks.last}.}

\item{numeric.type}{either "decimal" (default) or "roman".  Are numeric values represented as
    decimal numbers (\code{numeric.type="decimal"}) or as Roman numerals
    (\code{numeric.type="roman"})? }

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Functions sort selected lines. And sort numbers as numbers not alphabetically.
}
\note{
Inspired by \url{https://github.com/dcomtois/sortLines}.
}
