% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_align_code.R
\name{align_code}
\alias{align_code}
\alias{rs_align_code}
\alias{rs_align_code_at_equal}
\alias{rs_align_code_at_arrow}
\alias{rs_align_code_at_pipe_operator}
\alias{rs_align_code_at_hash}
\alias{rs_align_code_at_vertical_bar}
\title{Align code at certain symbols.}
\usage{
rs_align_code(at_symbol, context = rs_get_context())

rs_align_code_at_equal(context = rs_get_context())

rs_align_code_at_arrow(context = rs_get_context())

rs_align_code_at_pipe_operator(context = rs_get_context())

rs_align_code_at_hash(context = rs_get_context())

rs_align_code_at_vertical_bar(context = rs_get_context())
}
\arguments{
\item{at_symbol}{(character) \cr
A regular expression that represents symbol treated as reference for
code alignment.}

\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
In selected lines, align code at first occurance of certain symbol in each row.
}
\note{
Inspired by \url{https://github.com/seasmith/AlignAssign}.
}
