% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_arrows.R
\name{insert_arrows}
\alias{insert_arrows}
\alias{rs_insert_arrow_rl}
\alias{rs_insert_arrow_rl2}
\alias{rs_insert_arrow_lr}
\alias{rs_insert_arrow_lr2}
\title{Insert arrow (->, ->>, <-, <<-)}
\usage{
rs_insert_arrow_rl(context = rs_get_context())

rs_insert_arrow_rl2(context = rs_get_context())

rs_insert_arrow_lr(context = rs_get_context())

rs_insert_arrow_lr2(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
RStudio add-in to insert either \verb{->}, \verb{->>},
\verb{<-} or \verb{<<-} at the cursor position.
}
\note{
use hotkeys combination \kbd{ALT} + \kbd{-} to insert \verb{<-} in RStudio.
}
\examples{
\dontrun{
\donttest{
library(addins.rs)

rs_insert_arrow_rl()
## <-

rs_insert_arrow_rl2()
## <<-

rs_insert_arrow_lr()
## ->

rs_insert_arrow_lr2()
## ->>
}
}
}
\seealso{
Assignment operators \link[base:assignOps]{assignOps}.\cr

Other 'Insert at cursor position' addins: 
\code{\link{insert_line}}
}
\concept{'Insert at cursor position' addins}
