% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_infix_in}
\alias{rs_insert_infix_in}
\title{Insert \%in\%.}
\usage{
rs_insert_infix_in(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Call this function as an add-in to insert \code{\%in\%} at the cursor position.
}
\seealso{
Operator \link[base:match]{\%in\%}.

Other 'Insert at cursor position' add-ins: 
\code{\link{insert_magrittr_operators}},
\code{\link{rs_enclose_with_percent}()},
\code{\link{rs_insert_integer_division}()},
\code{\link{rs_insert_matrix_multiplication}()},
\code{\link{rs_insert_modulo}()}
}
\concept{'Insert at cursor position' add-ins}
