% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_integer_division}
\alias{rs_insert_integer_division}
\title{Insert \%/\%.}
\usage{
rs_insert_integer_division(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Call this function as an add-in to insert \code{ \%/\% } at the cursor position.
}
\seealso{
Operator \link[base]{\%/\%}.

Other 'Insert at cursor position' add-ins: 
\code{\link{insert_magrittr_operators}},
\code{\link{rs_enclose_with_percent}()},
\code{\link{rs_insert_infix_in}()},
\code{\link{rs_insert_matrix_multiplication}()},
\code{\link{rs_insert_modulo}()}
}
\concept{'Insert at cursor position' add-ins}
