% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_operators_other.R
\name{rs_insert_modulo}
\alias{rs_insert_modulo}
\title{Insert \%\%.}
\usage{
rs_insert_modulo(context = rs_get_context())
}
\arguments{
\item{context}{(class \code{document_context}) \cr
Object with context of active RStudio document.}
}
\description{
Call this function as an add-in to insert \code{\%\%} at the cursor position.
}
\note{
Operator does not work unless package which contains this operator is loaded.
}
\seealso{
Other 'Insert at cursor position' add-ins: 
\code{\link{insert_magrittr_operators}},
\code{\link{rs_enclose_with_percent}()},
\code{\link{rs_insert_infix_in}()},
\code{\link{rs_insert_integer_division}()},
\code{\link{rs_insert_matrix_multiplication}()}
}
\concept{'Insert at cursor position' add-ins}
